moved all "given" actions to "when"
